+++ app/assets/javascripts/activeadmin_addons/color-picker.js
@@ -2,7 +2,6 @@
 //= require_self
 
 $(function() {
-
   setupColorPicker();
 
   $(document).on('has_many_add:after', setupColorPicker);
@@ -10,7 +9,7 @@ $(function() {
   function setupColorPicker() {
     $('.color-picker').each(function(i, el) {
       $(el).paletteColorPicker({
-        clear_btn: 'last'
+        'clear_btn': 'last',
       });
     });
   }