+++ app/assets/javascripts/activeadmin_addons/selected-list.js
@@ -16,14 +16,15 @@ $(function() {
     $('.selected-list', container).each(function(i, el) {
       var url = $(el).data('url');
       var fields = $(el).data('fields');
-      var displayName = $(el).data('display_name');
+      var displayName = $(el).data('display-name');
       var width = $(el).data('width') || '100%';
       var method = $(el).data('method');
       var model = $(el).data('model');
       var prefix = model + '_' + method;
-      var responseRoot = $(el).data('response_root');
-      var minimumInputLength = $(el).data('minimum_input_length');
+      var responseRoot = $(el).data('response-root');
+      var minimumInputLength = $(el).data('minimum-input-length');
       var order = $(el).data('order') || (fields[0] + '_desc');
+
       var selectOptions = {
         width: width,
         minimumInputLength: minimumInputLength,